rename TransactionGatewayClient to TransactionClient
